Create README.md
Lead-Scoring-Case-Study
A logistic regression model to find potential leads of X Education

X Education sells online courses to industry professionals although X Education gets a lot of leads, its lead conversion rate is very poor. For example, if, say, they acquire 100 leads in a day, only about 30 of them are converted To make this process more efficient, the company wishes to identify the most potential leads, also known as ‘Hot Leads’ X Education has appointed you to help them select the most promising leads, i.e. the leads that are most likely to convert into paying customers. The company requires you to build a model wherein you need to assign a lead score to each of the leads such that the customers with higher lead score have a higher conversion chance and the customers with lower lead score have a lower conversion chance. The CEO, in particular, has given a ballpark of the target lead conversion rate to be around 80%.